maxDissimProgBar: Maximum Dissimilarity Sampling With A Progress Bar

This is the same function as in maxDissim but with the verbose argument replaced by a progress bar.

maxDissimProgBar(a, b, n = 2, obj = minDiss, useNames = FALSE,
  randomFrac = 1, ...)

Arguments

a
a matrix or data frame of samples to start

b
a matrix or data frame of samples to sample from

n
the size of the sub-sample

obj
an objective function to measure overall dissimilarity

useNames
a logical: should the function return the row names (as opposed ot the row index)

randomFrac
a number in (0, 1] that can be used to sub-sample from the remaining candidate values

...
optional arguments to pass to dist

Details

See vignette("vis2rfi") for a tutorial/guide on how the package can be used.

See also